# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.CategoryInput:
    properties:
      description:
        type: string
      name:
        type: string
      parent_id:
        type: integer
    type: object
...
  handlers.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  handlers.TransactionInput:
    properties:
      amount:
        type: number
      category_id:
        type: integer
      date:
        description: Expects "2006-01-02"
        type: string
      description:
        type: string
      receipt_image_url:
        type: string
      sub_category_id:
        type: integer
      type:
        type: string
      user_id:
        type: integer
      wallet_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@expense-tracker.local
    name: API Support
    url: http://www.swagger.io/support
  description: Expense Tracker REST API for users, categories, transactions, balances,
    wallets.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Expense Tracker API
  version: "1.0"
paths:
...
  /api/categories:
    get:
      description: Get top-level categories (parent_id IS NULL) (protected)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: List categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create a new category or subcategory (protected)
      parameters:
      - description: Category create request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CategoryInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Create category
      tags:
      - Categories
  /api/categories/{id}:
    delete:
      description: Delete category by id (protected)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Delete category
      tags:
      - Categories
    get:
      description: Get single category by id (protected)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Get category
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Update category metadata (protected)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Category update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CategoryInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Update category
      tags:
      - Categories
  /api/categories/{id}/sub-categories:
    get:
      description: Get subcategories of a category (protected)
      parameters:
      - description: Parent Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: List sub-categories
      tags:
      - Categories
  /api/dashboard/stats:
    get:
      description: Get transaction dashboard stats for a user (protected)
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Get dashboard stats
      tags:
      - Transactions
  /api/transactions:
    get:
      description: Get all transactions for a user (protected)
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Get transactions
      tags:
      - Transactions
    post:
      consumes:
      - application/json
      description: Create a new transaction record (protected)
      parameters:
      - description: Transaction create request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.TransactionInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Create transaction
      tags:
      - Transactions
  /api/transactions/{id}:
    delete:
      description: Delete transaction by id for a user (protected)
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Delete transaction
      tags:
      - Transactions
    get:
      description: Get transaction by id for a user (protected)
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Get transaction
      tags:
      - Transactions
    put:
      consumes:
      - application/json
      description: Update transaction record by id (protected)
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      - description: Transaction update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.TransactionInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Update transaction
      tags:
      - Transactions
...
schemes:
- http
swagger: "2.0"